Update workflows, Docker setup, and internal structures; remove tracking for SSL certs

+++ internal/models/document_ownership.go
@@ -13,8 +13,7 @@ type DocumentOwnership struct {
 }
 
 func (d *DocumentOwnership) Insert(pool *pgxpool.Pool, resultChan chan<- ResultChan[string]) {
-	query := `INSERT INTO public.document_ownerships (user_id, document_id) VALUES ($1, $2)`
-
+	query := `INSERT INTO public.document_ownerships (user_id, document_id) VALUES ($1, $2) RETURNING document_id`
 	var docID string
 	err := pool.QueryRow(context.Background(), query, d.UserID, d.DocumentID).Scan(&docID)
 	if err != nil {
@@ -31,15 +30,15 @@ func (d *DocumentOwnership) Update(pool *pgxpool.Pool, resultChan chan<- ResultC
 }
 
 func (d *DocumentOwnership) Delete(pool *pgxpool.Pool, resultChan chan<- ResultChan[error]) {
-	query := `DELETE FROM public.document_ownerships WHERE document_id = $1`
-	_, err := pool.Exec(context.Background(), query, d.DocumentID)
+	query := `DELETE FROM public.document_ownerships WHERE document_id = $1 AND user_id = $2`
+	_, err := pool.Exec(context.Background(), query, d.DocumentID, d.UserID)
 	resultChan <- ResultChan[error]{Error: err}
 }
 
 func (d *DocumentOwnership) Query(pool *pgxpool.Pool, resultChan chan<- ResultChan[*DocumentOwnership]) {
 	var doc DocumentOwnership
-	query := `SELECT user_id, document_id FROM public.document_ownerships WHERE document_id = $1`
-	err := pool.QueryRow(context.Background(), query, d.DocumentID).Scan(&doc.UserID, &doc.DocumentID)
+	query := `SELECT user_id, document_id FROM public.document_ownerships WHERE document_id = $1 AND user_id = $2`
+	err := pool.QueryRow(context.Background(), query, d.DocumentID, d.UserID).Scan(&doc.UserID, &doc.DocumentID)
 	if err != nil {
 		resultChan <- ResultChan[*DocumentOwnership]{Error: err}
 		return
